backend-ml/requirements.txt
fastapi==0.104.0
uvicorn[standard]==0.23.2
pydantic==2.4.2
scikit-learn==1.3.1
pandas==2.1.1
numpy==1.26.0
geopy==2.4.0
google-generativeai==0.5.2
redis==5.0.4
httpx==0.27.0
python-dotenv==1.0.1
